Introduce support for building on MinGW/Cygwin systems
This patch modifies the build to properly handle MinGW/Cygwin
hosts and adds support for Windows sockets in the source.  This was
achieved by using MHD_socket as the socket datatype where necessary,
which is guaranteed to resolve to SOCKET on windows and int on unix
systems, preserving old functionality while adding new.  Some internal
functions were modified to accomodate the change, namely those that
implictly converted sockets to ints (which would have caused problems on
x86_64 windows, as those are 8 byte unsigned) All test cases
pass with MSYS2/MinGW compiler.

+++ src/http_response.cpp
@@ -207,7 +207,7 @@ namespace details
 
 ssize_t cb(void* cls, uint64_t pos, char* buf, size_t max)
 {
-    int val = static_cast<http_response*>(cls)->cycle_callback(buf);
+    ssize_t val = static_cast<http_response*>(cls)->cycle_callback(buf);
     if(val == -1)
         static_cast<http_response*>(cls)->completed = true;
     return val;
@@ -272,7 +272,7 @@ ssize_t http_response::data_generator(
     if(_this->ws->pop_signaled(_this->connection_id))
     {
         string message;
-        int size = _this->ws->read_message(_this->connection_id, message);
+        size_t size = _this->ws->read_message(_this->connection_id, message);
         memcpy(buf, message.c_str(), size);
         return size;
     }
